backend/requirements.txt
Flask
Flask-CORS
SQLAlchemy
Werkzeug
mysql-connector-python
alembic
python-dotenv
bcrypt
PyPDF2
psycopg2-binary
pyodbc